error handling

+++ app/Http/Controllers/CategoryController.php
@@ -94,6 +94,12 @@ class CategoryController extends Controller
      */
     public function destroy(Category $category)
     {
+        if ($category->posts->count() > 0) {
+            session()->flash('error', 'This category still has posts in it!');
+
+            return redirect()->back();
+        }
+
         $category->delete();
 
         session()->flash('success', 'Category successfully Deleted!');
